feat: run using composite (#3)
* feat: run using `composite`

The action no longer runs in a docker container, cutting down the setup
time. Before this change the setup of this action took approximately 15
seconds and is now around 1 to 3 seconds.

This means that this action will only work in actions that runs on
linux/ ubuntu.

* fix: use `uuidgen` for request id

The previous method for getting the random request id was not able to
run outside of the docker container.